Epurse
Block functionality in the card.

The personalization terminal stores in the card the month and the year that the card will expire.
Also, the personalization terminal update a file with the ID of the cards.

We suppose that this file is in the online system.

Also, there is a block_id.txt file, that contains the id of the cards that has to be blocked, for example in case that the user lost his card.

The EpurseTerminal does three checks. Check if the card is ISSUED. If yes, it checks if the card's id is in the block_list file. If yes, blocks the card. Otherwise, checks the exparation date.